+++ js/main.js
@@ -1,8 +1,7 @@
-import {createPhotos} from './modules/data.js';
 import './modules/file-upload.js';
 import './modules/photo-scale-control.js';
 import './modules/photo-upload-form.js';
-import './modules/picture-list.js';
+import {createPhotos} from './modules/picture-list.js';
 import './modules/effects-control.js';
 
 createPhotos();
